requirements.txt
PySide6==6.10.1
SQLAlchemy==2.0.32
psycopg[binary]==3.2.13
python-dotenv==1.0.1
alembic==1.13.2
pytest==8.3.2
